tg-bot/requirements.txt
python-telegram-bot
SQLAlchemy
psycopg2
sqlacodegen
pygsheets
